requirements.txt
marshmallow
django
marshmallow-jsonschema
djangorestframework-jsonapi
djangorestframework
wrapt
jsonfield
